requirements.txt
opencv-python
numpy
pillow
pixels2svg
moms_apriltag
